+++ src/apis/auth.ts
@@ -1,10 +1,9 @@
-import { headers } from 'next/dist/client/components/headers';
 import client from './client';
 
 export const register = async (input: any) => {
   const response = await client.post('/auth/signup/', {
     name: input.name,
-    id: input.id,
+    login_id: input.id,
     password: input.password,
     email: input.email,
     part: input.part,
@@ -13,7 +12,6 @@ export const register = async (input: any) => {
   return response.data;
 };
 
-//login
 export const login = async (input: any) => {
   const response = await client.post('/auth/signin/', {
     login_id: input.id,
@@ -22,27 +20,10 @@ export const login = async (input: any) => {
   return response.data;
 };
 
-//refreshToken, accessToken 재발급
-export const mutateRefreshing = async () => {
-  return client.post('/auth/token/refresh/').then(response => {
-    return response.data;
-  });
-};
-
-//logout
-export const logout = async (accessToken: string) => {
-  const response = await client.post('/auth/signout/', {
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  });
-  return response.data;
-};
-
 //id,email 중복 확인
 export const checkId = async (id: string) => {
   const response = await client.post('/auth/id/check/', {
-    id: id,
+    login_id: id,
   });
   return response.data;
 };
@@ -54,9 +35,8 @@ export const checkEmail = async (email: string) => {
   return response.data;
 };
 
-//데모데이 투표 중복 제한
-export const demoDayAuthority = async (accessToken: string) => {
-  const response = await client.post('/votes/teams/authority/', {
+export const logout = async (accessToken: string) => {
+  const response = await client.post('/auth/signout/', {
     headers: {
       Authorization: `Bearer ${accessToken}`,
     },
@@ -64,24 +44,7 @@ export const demoDayAuthority = async (accessToken: string) => {
   return response.data;
 };
 
-//데모데이 투표
-export const demoDayVote = async (info: any) => {
-  const response = await client.post(
-    '/votes/teams/',
-    {
-      tname: info.tname,
-    },
-    {
-      headers: {
-        Authorization: `Bearer ${info.accessToken}`,
-      },
-    },
-  );
-  return response.data;
-};
-
-//데모데이 결과
-export const getDemoDayResult = async () => {
-  const response = await client.get('/votes/teams/result/');
+export const tokenRefresh = async () => {
+  const response = await client.post('/auth/token/refresh/');
   return response.data;
 };
